Repair nav item highlighting in main menu
It appears that the `250` passed as a parameter to the scroll
event was causing some strange behavior. Removing that ultimately
fixed the nav item highlighting issue.

In the process, I removed the extra code used to assign `dogeared`
to nav items and turned off the debounce feature in doormat, which
shouldn't be in play now that we have snapping disabled.

Fixes #25.
Fixes #21.