Use `the-evidence`, `spine-mobile`, and `spine-full` classes for targeting
Using `spine-mobile` and `spine-full` reduces the dependency on media queries.
The `spine-mobile` class is leveraged to modify the break points in preparation
for mobile background images for columns.
Adding `the-evidence` as a body class to the markup and using it for targeting
elements should prevent the styles from conflicting with any current pages and
custom styles.